requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-celery-results==2.5.1
django-leaflet==0.29.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
kombu==5.3.4
prompt-toolkit==3.0.43
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
